+++ exec/Kathu/App/Tools/ToolMode.hs
@@ -34,6 +34,10 @@ shouldShowGrid :: ToolMode -> Bool
 shouldShowGrid NoTool = False
 shouldShowGrid _      = True
 
+canEditEntities :: ToolMode -> Bool
+canEditEntities NoTool = False
+canEditEntities _      = True
+
 -- | Returns True if both modes are the same type, regardless of configuration
 isSameMode :: ToolMode -> ToolMode -> Bool
 isSameMode NoTool       NoTool       = True
@@ -42,11 +46,12 @@ isSameMode SignalWirer  SignalWirer  = True
 isSameMode _            _            = False
 
 data ToolModeUniversalState = ToolModeUniversalState
-    { lastPlacedTilePos :: !(Maybe (V2 Int))
-    , selectedTile      :: !(Tile ImageID)
-    , lastUndoRedoTime  :: !Word32
+    { lastPlacedTilePos     :: !(Maybe (V2 Int))
+    , selectedTile          :: !(Tile ImageID)
+    , lastUndoRedoTime      :: !Word32
+    , canEditEntityInstance :: !Bool
     }
 
 instance Semigroup ToolModeUniversalState where (<>) = mappend
-instance Monoid ToolModeUniversalState where mempty = ToolModeUniversalState Nothing emptyTile 0
+instance Monoid ToolModeUniversalState where mempty = ToolModeUniversalState Nothing emptyTile 0 True
 instance Component ToolModeUniversalState where type Storage ToolModeUniversalState = Global ToolModeUniversalState