+++ exec/Kathu/Editor/Util/GtkMisc.hs
@@ -2,11 +2,12 @@
 
 module Kathu.Editor.Util.GtkMisc where
 
-import           Control.Monad              (forM_)
+import           Control.Monad              (forM_, void, when)
 import           Data.GI.Base
-import           Data.GI.Base.GType         (gtypeStrv)
+import           Data.GI.Base.GType         (gtypeString)
 import           Data.Text                  (Text)
 import qualified Data.Text                  as T
+import qualified GI.GObject.Functions       as GObj
 import qualified GI.Gtk                     as Gtk
 
 import           Kathu.Editor.Types
@@ -33,12 +34,21 @@ mkMenuBar menuBarDesc = do
 
 mkEntryCompletion :: Foldable t => t Text -> IO Gtk.EntryCompletion
 mkEntryCompletion rows = do
-    list <- Gtk.listStoreNew [gtypeStrv]
+    list <- Gtk.listStoreNew [gtypeString]
     forM_ rows $ \row -> do
         txtVal  <- toGValue (Just row)
         newIter <- Gtk.listStoreAppend list
         Gtk.listStoreSetValue list newIter 0 txtVal
-    new Gtk.EntryCompletion [#model := list]
+    entryCompletion <- new Gtk.EntryCompletion [#model := list]
+    Gtk.entryCompletionSetTextColumn entryCompletion 0
+    pure entryCompletion
+
+showErrorDialog :: Text -> IO ()
+showErrorDialog msg = do
+    dialog <- new Gtk.MessageDialog [#messageType := Gtk.MessageTypeError, #text := msg]
+    void $ Gtk.dialogAddButton dialog "Accept" 1
+    void $ Gtk.dialogRun dialog
+    Gtk.widgetDestroy dialog
 
 showFileChooser :: Gtk.ApplicationWindow -> Gtk.FileChooserAction -> Text -> Text -> Text -> IO (Maybe String)
 showFileChooser window action title filterName filterExt = do
@@ -55,3 +65,28 @@ showFileChooser window action title filterName filterExt = do
     case toEnum (fromIntegral response) of
         Gtk.ResponseTypeAccept -> Gtk.fileChooserGetFilename dialog
         _                      -> pure Nothing
+
+-- | Contructs an Entry that refuses to allow text that cannot be parsed using the given parser
+mkParsableEntry :: (Text -> Maybe a) -> IO Gtk.Entry
+mkParsableEntry tryParse = do
+    entry       <- new Gtk.Entry []
+    -- kills a signal and prevents changes the the entry's text when it fails to parse
+    let cancelInvalid sigName newTxt = case tryParse newTxt of
+            Just _  -> pure ()
+            Nothing -> GObj.signalStopEmissionByName entry sigName
+
+    void $ on entry #insertText $ \newTextPart _ pos -> do
+        prevText <- get entry #text
+        let (beginTxt, endTxt) = T.splitAt (fromIntegral pos) prevText
+            newTxt             = T.concat [beginTxt, newTextPart, endTxt]
+        cancelInvalid "insert-text" newTxt
+        pure pos
+
+    void $ on entry #deleteText $ \startPos endPos ->
+        when (startPos >= 0 && endPos >= 0) $ do
+            prevText <- get entry #text
+            let beginTxt = T.take (fromIntegral startPos) prevText
+                endTxt   = T.drop (fromIntegral endPos)   prevText
+                newTxt   = beginTxt `T.append` endTxt
+            cancelInvalid "delete-text" newTxt
+    pure entry